can't be separate files if allocatable is used as inout intent, needs to be in interface or contains block
https://stackoverflow.com/questions/48472442/can-i-use-allocatable-array-as-an-intentout-matrix-in-fortran

It is perfectly fine to pass an allocatable dummy argument to subroutine in Fortran 2003 and later. IIRC it was allowed by a TS (technical specification) to Fortran 95 first. All compilers support it today.
The actual argument does not have to be allocated when being passed to an allocatable dummy argument. Here it will actuully be automatically deallocated if it happens to be allocated, because the dummy is intent(out) and allocatable, intent(out) arguments are automatically deallocated when entering the subroutine.
Your code should work fine, but there is one important aspect, the subroutine must have an explicit interface. So it must be placed in a module. Or it must be internal (after contains) or an interface block must be used (ugly).